# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Nakudin Marketplace API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /products:
    get:
      operationId: listProducts
      tags: [products]
      summary: List products with optional filters
      parameters:
        - in: query
          name: shopId
          schema: { type: ["string", "null"] }
        - in: query
          name: category
          schema: { type: ["string", "null"] }
        - in: query
          name: cursor
          schema: { type: ["string", "null"] }
        - in: query
          name: limit
          schema: { type: integer, default: 20 }
      responses:
        "200":
          description: Product list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductPage"

    post:
      operationId: createProduct
      tags: [products]
      summary: Create a new product
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
  /products/{productId}:
    get:
      operationId: getProduct
      tags: [products]
      summary: Get a product by ID
      parameters:
        - in: path
          name: productId
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"

    patch:
      operationId: updateProduct
      tags: [products]
      summary: Update a product
      parameters:
        - in: path
          name: productId
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"

    delete:
      operationId: deleteProduct
      tags: [products]
      summary: Delete a product
      parameters:
        - in: path
          name: productId
          required: true
          schema: { type: string }
      responses:
        "204":
          description: Deleted
  /products/{productId}/like:
    post:
      operationId: likeProduct
      tags: [social]
      summary: Like a product
      parameters:
        - in: path
          name: productId
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Liked
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LikeResult"

    delete:
      operationId: unlikeProduct
      tags: [social]
      summary: Unlike a product
      parameters:
        - in: path
          name: productId
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Unliked
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LikeResult"
  /products/{productId}/comments:
    get:
      operationId: getProductComments
      tags: [social]
      summary: Get comments for a product
      parameters:
        - in: path
          name: productId
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Comments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Comment"

    post:
      operationId: createComment
      tags: [social]
      summary: Post a comment on a product
      parameters:
        - in: path
          name: productId
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CommentInput"
      responses:
        "201":
          description: Comment created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Comment"
  /products/{productId}/whatsapp-click:
    post:
      operationId: logWhatsappClick
      tags: [social]
      summary: Log a WhatsApp button click
      parameters:
        - in: path
          name: productId
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WhatsappClickInput"
      responses:
        "200":
          description: Logged
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WhatsappClickResult"
  /products/{productId}/related:
    get:
      operationId: getRelatedProducts
      tags: [products]
      summary: Get related products
      parameters:
        - in: path
          name: productId
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Related products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/FeedProduct"
...
components:
  schemas:
...
    FeedProduct:
      type: object
      required: [id, title, price, images, likeCount, viewCount, shopId, shopName, shopVerified, createdAt]
      properties:
        id: { type: string }
        title: { type: string }
        price: { type: number }
        images: { type: array, items: { type: string } }
        likeCount: { type: integer }
        viewCount: { type: integer }
        shopId: { type: string }
        shopName: { type: string }
        shopVerified: { type: boolean }
        shopLogoUrl: { type: ["string", "null"] }
        shopWhatsapp: { type: ["string", "null"] }
        category: { type: ["string", "null"] }
        locationCity: { type: ["string", "null"] }
        locationState: { type: ["string", "null"] }
        distanceKm: { type: ["number", "null"] }
        isLiked: { type: boolean }
        isFollowed: { type: boolean }
        trendScore: { type: ["number", "null"] }
        createdAt: { type: string }
...
    Product:
      type: object
      required: [id, shopId, title, price, images, likeCount, viewCount, whatsappClickCount, status, createdAt]
      properties:
        id: { type: string }
        shopId: { type: string }
        title: { type: string }
        description: { type: ["string", "null"] }
        price: { type: number }
        images: { type: array, items: { type: string } }
        category: { type: ["string", "null"] }
        locationCity: { type: ["string", "null"] }
        locationState: { type: ["string", "null"] }
        locationLat: { type: ["number", "null"] }
        locationLng: { type: ["number", "null"] }
        likeCount: { type: integer }
        viewCount: { type: integer }
        whatsappClickCount: { type: integer }
        trendScore: { type: ["number", "null"] }
        status: { type: string, enum: [active, hidden, deleted] }
        isLiked: { type: boolean }
        createdAt: { type: string }
    ProductInput:
      type: object
      required: [title, price, images]
      properties:
        title: { type: string, minLength: 1 }
        description: { type: string }
        price: { type: number, minimum: 0 }
        images: { type: array, items: { type: string }, maxItems: 6 }
        category: { type: string }
        locationCity: { type: string }
        locationState: { type: string }
        locationLat: { type: number }
        locationLng: { type: number }
    ProductUpdate:
      type: object
      properties:
        title: { type: string, minLength: 1 }
        description: { type: string }
        price: { type: number, minimum: 0 }
        images: { type: array, items: { type: string }, maxItems: 6 }
        category: { type: string }
        status: { type: string, enum: [active, hidden] }
        locationCity: { type: string }
        locationState: { type: string }
    ProductPage:
      type: object
      required: [products, nextCursor]
      properties:
        products:
          type: array
          items:
            $ref: "#/components/schemas/Product"
        nextCursor: { type: ["string", "null"] }
...
    LikeResult:
      type: object
      required: [liked, likeCount]
      properties:
        liked: { type: boolean }
        likeCount: { type: integer }
...
    Comment:
      type: object
      required: [id, productId, authorId, authorName, text, createdAt]
      properties:
        id: { type: string }
        productId: { type: string }
        authorId: { type: string }
        authorName: { type: string }
        authorLogoUrl: { type: ["string", "null"] }
        text: { type: string }
        createdAt: { type: string }
    CommentInput:
      type: object
      required: [text]
      properties:
        text: { type: string, minLength: 1 }
...
    WhatsappClickInput:
      type: object
      required: [deviceId]
      properties:
        deviceId: { type: string }
    WhatsappClickResult:
      type: object
      required: [logged]
      properties:
        logged: { type: boolean }
...