# lib/api-zod/src/generated/types/comment.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Nakudin Marketplace API
 * OpenAPI spec version: 0.1.0
 */

export interface Comment {
  id: string;
  productId: string;
  authorId: string;
  authorName: string;
  /** @nullable */
  authorLogoUrl?: string | null;
  text: string;
  createdAt: string;
}