# lib/api-zod/src/generated/types/likeResult.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Nakudin Marketplace API
 * OpenAPI spec version: 0.1.0
 */

export interface LikeResult {
  liked: boolean;
  likeCount: number;
}