# lib/api-zod/src/generated/types/productUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Nakudin Marketplace API
 * OpenAPI spec version: 0.1.0
 */
import type { ProductUpdateStatus } from './productUpdateStatus';

export interface ProductUpdate {
  /** @minLength 1 */
  title?: string;
  description?: string;
  /** @minimum 0 */
  price?: number;
  /** @maxItems 6 */
  images?: string[];
  category?: string;
  status?: ProductUpdateStatus;
  locationCity?: string;
  locationState?: string;
}